add preprocess options on message level
