+++ example/proto/demo.pb.preprocess.go
@@ -17,8 +17,8 @@ var _ = math.Inf
 
 func (m *Demo) Preprocess() error {
 
-	m.PreprocessedField = strings.TrimSpace(m.PreprocessedField)
 	m.PreprocessedField = strings.ToLower(m.PreprocessedField)
+	m.PreprocessedField = strings.TrimSpace(m.PreprocessedField)
 
 	for i, s := range m.PreprocessedRepeatedField {
 		m.PreprocessedRepeatedField[i] = strings.TrimSpace(s)
