+++ options/preprocess.proto
@@ -5,13 +5,22 @@ option go_package = "preprocess";
 
 import "google/protobuf/descriptor.proto";
 
-// Field level specifications
+// Message level options
+extend google.protobuf.MessageOptions{
+    PreprocessMessageOptions each = 11111;
+}
+
+message PreprocessMessageOptions {
+    PreprocessString string = 1;
+}
+
+// Field level options
 extend google.protobuf.FieldOptions {
     PreprocessFieldOptions field = 11111;
 }
 
 message PreprocessFieldOptions {
-    PreprocessString string = 2;
+    PreprocessString string = 1;
 }
 
 message PreprocessString {
