+++ preprocess/preprocess.go
@@ -1,11 +1,13 @@
 package plugin
 
 import (
-	"github.com/gogo/protobuf/proto"
-	"github.com/gogo/protobuf/protoc-gen-gogo/descriptor"
 	"github.com/gogo/protobuf/protoc-gen-gogo/generator"
 	prep "github.com/infobloxopen/protoc-gen-preprocess/options"
 )
+
+func init() {
+	generator.RegisterPlugin(NewPreprocessor())
+}
 
 type preprocessor struct {
 	*generator.Generator
@@ -31,32 +33,14 @@ func (p *preprocessor) Generate(file *generator.FileDescriptor) {
 	p.PluginImports = generator.NewPluginImports(p.Generator)
 	p.stringsPkg = p.NewImport("strings")
 	for _, message := range file.Messages() {
-		if containsPreprocessOptions(message) {
-			p.generateProto3Message(message)
+		opts := getMessageOptions(message)
+		if opts != nil || containsFieldPreprocessOptions(message) {
+			p.generateProto3Message(message, opts)
 		}
 	}
 }
 
-func init() {
-	generator.RegisterPlugin(NewPreprocessor())
-}
-
-func getFieldOptions(field *descriptor.FieldDescriptorProto) *prep.PreprocessFieldOptions {
-	if field.Options == nil {
-		return nil
-	}
-	v, err := proto.GetExtension(field.Options, prep.E_Field)
-	if err != nil {
-		return nil
-	}
-	opts, ok := v.(*prep.PreprocessFieldOptions)
-	if !ok {
-		return nil
-	}
-	return opts
-}
-
-func (p *preprocessor) generateProto3Message(message *generator.Descriptor) {
+func (p *preprocessor) generateProto3Message(message *generator.Descriptor, messageOptions *prep.PreprocessMessageOptions) {
 	ccTypeName := generator.CamelCaseSlice(message.TypeName())
 	p.P(`func (m *`, ccTypeName, `) Preprocess() error {`)
 	p.In()
@@ -67,9 +51,8 @@ func (p *preprocessor) generateProto3Message(message *generator.Descriptor) {
 		}
 		fieldName := p.GetOneOfFieldName(message, field)
 		variableName := "m." + fieldName
-		repeated := field.IsRepeated()
 		if field.IsString() {
-			p.generateStringPreprocessor(variableName, fieldOptions, repeated)
+			p.generateStringPreprocessor(variableName, []prepOptions{messageOptions, fieldOptions}, field.IsRepeated())
 		}
 	}
 	p.Out()
@@ -79,36 +62,27 @@ func (p *preprocessor) generateProto3Message(message *generator.Descriptor) {
 	p.P()
 }
 
-var stringMethods = map[prep.PreprocessString_Methods]string{
-	prep.PreprocessString_trim_space: ".TrimSpace",
-	prep.PreprocessString_upper:      ".ToUpper",
-	prep.PreprocessString_lower:      ".ToLower",
-}
-
-func (p *preprocessor) generateStringPreprocessor(variableName string, fv *prep.PreprocessFieldOptions, repeated bool) {
-	str := fv.GetString_()
+func (p *preprocessor) generateStringPreprocessor(variableName string, opts []prepOptions, repeated bool) {
 	p.P()
+	strMethods := make(map[string]prep.PreprocessString_Methods)
+	for _, v := range opts {
+		if str := v.GetString_(); str != nil {
+			for _, m := range str.Methods {
+				strMethods[m.String()] = m
+			}
+		}
+	}
 	if repeated {
 		p.P(`for i, s := range `, variableName, `{`)
 		p.In()
-		for _, method := range str.Methods {
+		for _, method := range strMethods {
 			p.P(variableName, `[i] = `, p.stringsPkg.Use(), stringMethods[method], `(s)`)
 		}
 		p.Out()
 		p.P(`}`)
 	} else {
-		for _, method := range str.Methods {
+		for _, method := range strMethods {
 			p.P(variableName, ` = `, p.stringsPkg.Use(), stringMethods[method], `(`, variableName, `)`)
 		}
 	}
-}
-
-func containsPreprocessOptions(message *generator.Descriptor) bool {
-	for _, field := range message.Field {
-		fieldOptions := getFieldOptions(field)
-		if fieldOptions != nil {
-			return true
-		}
-	}
-	return false
 }
